refactor(tests): Move server parameter tests to new `ServerParamsPsr7Test` class for better organization and clarity.
